Forward the server error response to the client if possible
If the Http::Tunneler  received a valid  error response from peer forward it
to the client instead of sending a squid generated error page.

This patch add the ERR_RELAY_REMOTE err_type to describe ErrorState objects
which are created over a peer HttpReply.

Maybe the Http::Tunneler should wait to retrieve all of body of peer response.
Currently waits and forwards only the HTTP headers.