Removed TODO based on Christos's investigation of current practices
... and complications that make correct timeout canceling difficult.

Passing invalid timeouts to others is essentially a bug, but lots of
code does that, most bugs are usually harmless, and fixing all those
bugs is quite difficult (because there is no API for the cleanup code to
know whether the currently set timeout belongs to this job). Tunneler
is not the best place to bring attention to this complex problem.