Polished resurrected PROXYPASS documentation
Resurrecting PROXYPASS documentation from squid.conf may have been a
good idea, but that squid.conf documentation covered _all_ PROXYPASS
cases while the if-statement where it was resurrected covered a very
specific/special sub-case. Placing the original documentation here "as
is" resulted in comments that did not match the code and confused me
quite a bit! I adjusted the docs to match their location and to
emphasize that this is not the only place where PROXYPASS is handled.

Also split the long/complex if-statement condition into multiple lines.