Fix Http::Tunneler CONNECT related issues with acls
When Http::Tunneler is used with stare/peek/server-first bumping modes then
it is actually forwards a CONNECT request to the upstream proxy.
When the client-first (bump on step1) bumping mode is used, the squid
forwarding code actually receives a "GET https://..." request to forward.
The Http::Tunneler uses this GET request in acl checks while building
the outgoing HttpHeaders, causes acl mismatch for example when the
"method" acl is used.

This patch builds and uses a new HttpRequest object based on CONNECT method
to use with these acl checks.