Do not let keepForLocalMemoryCache() make keepInLocalMemory decision
... when there is a shared memStore.

MemStore should decide whether to cache the entry in shared memory.
Regardless of that decision, the entry should not be cached locally (to
reduce conflicts between shared and local memory caches/indexes).

Also used a more simple/direct "ban transient readers" logic instead of
"allow non-transients and transient writers" logic. Note that ioDone
entries (from xitTable.io point of view) are non-transients so both this
and the previous branch commit allow them. This may be wrong (we should
not (re)store something in a shared memory cache without locking
transients), but that is a different problem. We are trying to fix
the ioWriting assertion here.

This change reworks the previous branch commit logic.