Properly handle short responses with arbitrary bytes
Before this change a response having <4 characters was converted by
Squid into an invalid 'HTTP/1.1 0 Init" response with these characters
as body bytes. In some cases, e.g., with ICAP respmod adaptation
configured, the body was not sent to client at all, because ICAP
echo mode could not acquire it.

I fixed Squid to handle such response similarly as it already handles
responses with >=4 arbitrary bytes, i.e., convert it into a valid 200
response (with 'X-Transformed-From: HTTP/0.9' header and the bytes as
body.