Abort on listen(2) error
If the first Squid instance is running with worker-queues option
configured for a port, the second Squid started on the same port
but without worker-queues option reports listen(2) error. The problem
is that despite the error, Squid keeps running even without listening
sockets and confusing the user with incorrect "Accepting HTTP Socket
connections at local..." message.
This is a short-term solution, treating each listen() error as fatal.
A proper long-term solution would be to abort only if listen()
failed for all configured ports (just as bind(2) errors are dealt with).