Unexpected ERR_CANNOT_FORWARD while connecting to remote TLS servers
... if the servers contain huge certificates, eg while connecting to
1000-sans.badssl.com.
In this cases the server handshake messages (which include server certificates)
may sent to more than one TLS records. If squid reads these TLS records in
one read, it fail to parse them correctly and the connection freezes in
squid side waiting for more server TLS data.

This is a Measurement Factory project