Disabled on-disk partial writes in rock
These writes were re-enabled in 5296bb, which claimed that on-disk
collapsing should work. This collapsing (as well as the shared memory
collapsing) requires the 'read-while-writing' feature, that is
what StoreMap::startAppending() is responsible for.  However, it proved
to be that the ability to read from disk during writing has never been
fully implemented (it is known that just enabling it via
startAppending() is not sufficient). So, for performance sake,
it is better to disable (yet useless) partial writes.

Except this performance issue, there is another problem caused by
93bdc39. droppedEarlierRequest() method, added at that revision, is
incapable of dealing with partial writes, returning an error.  As a
result, the whole entry is marked for deletion, and already collapsed
entries are aborted. Obviously, disabling partial writes resolves
this problem as well.