Make the ConnStateData::borrowPinnedConnection to throw on errors
Also the ConnStateData::validatePinnedConnection now throws on error
and converted to a void method.
Both methods throws a PinningException exception which is a
std::runtime_error kid and also has a code for the type of error:
  * errPolicy error which means the pinned connection policy violated
  * errConnectionGone if one of the pinned connection sides is closed

The new static ConnStateData::BorrowPinnedConnection method is added
to check also for valid ConnStateData object.