Removed ENTRY_REQUIRES_COLLAPSING flag
Thus addressing 325afae TODO. Though we intended first to use nil
MemObject::reply_ as a token of CF risks, it proved to be fairly
difficult do follow this approach because MemObject::reply_ is always
non-nil since it is created in MemObject constructor.  I considered
refactoring so that MemObject::reply_ was initialized only when a reply
is received but rejected this idea because MemObject::reply_ is already
used in many contexts expecting it non-nil and it is not possible easy
to circumvent this (e.g., by means of some other MemObject field)
without large out-of-scope refactoring.  Instead, we can rely on the
assumption that the received reply headers mean the end of
collapsing for this entry. I used Http::Message::pstate for this
purpose.