Added AsyncContext class
This class saves/restores transaction context. The context is
used in debugs() level 0,1 messages. In order to provide the context
information to debugs() in various async contexts, before any
scheduling of async call or SetSelect, the transaction context
should be saved and at the beginning of the async call or DoSelect()
restored. For this purpose, the data objects for async calls
and SetSelect need to reimplement AsyncContext::context().

This is a kind of research of possible ways of debugging
transaction contexts rather than a final solution.