Store/restore transaction context in DNS queries
While the query itself is performed on behalf on a single
transaction, other transactions may be waiting and
synchronously receive the reply. It looks like that the loop
over all receients is performed via
idnsCallbackAllCallersWithNewAnswer()/idnsCallbackOneWithAnswer(),
while PeerSelector::selectMore() is the initiator and
PeerSelector::handlePath() the receiver of the query, where
the async context should be stored/restored accordingly.