Allow ConnOpener handle failed connections to peers
The handling is performed via peerConnectFailed(). Instead of doing this
work in each caller (usually in a callback passed to ConnOpener job), it
is better to avoid this duplication and call the method from ConnOpener
itself. Note that Security::BlindPeerConnector already implements this
approach.

Also do not forget to delete ErrorState object: now
TunnelsStateData::retry() takes care about it.