The server_cert_fingerprint acl does not match when it is used with the sslproxy_cert_error access list.
The issue appears when the certificate validator is used. In
this case the certificate is not passed to the related
ACLChecklist object and it is not available for acl checks
where it is required.

This is a Measurement Factory project

Applied SQUID-462-Fix-server_cert_fingerprint-t1.patch
or 82e738c688b1702e7ce55003845092598eef57a7.