Write on disk the updated StoreEntry::swap_file_sz
This field, unlike Http headers, cannot be updated while processing 304
response: it is calculated as a sum of serialized data while writing the
entry on disk. Before this fix, the updated swap_file_sz was not stored
on disk because we wrote disk metadata (including stale swap_file_sz)
first, but calculated fresh swap_file_sz in the end, after the body
prefix was written. Fortunately, it is possible to calculate
this field before writing metadata with little overhead: we need to
serialize (and drop) meta header in order to get its size.