Use scoped enum for transaction steps
Scoped enums are generally better, but there was a specific trigger for
this change as well: Unscoped enum cannot have enumBegin_/enumEnd_
members because they are going to clash with same-name members of other
unscoped enums. And these members are required to use WholeEnum-based
iteration.

static_cast<>s are required for array sizes/indexes because scope enum
members are not considered to be of integral type in these expressions.
This is probably a side effect of prohibiting silent conversion between
the scoped enum members and integers.