Ignore orphaned Transients entries
This problem was discovered for rock-only cache configurations.

After a reply has been cached on disk and the transaction finishes,
the corresponding Transients entry is not freed. Another request
(with a different public key) may occupy the same rock slot (overriding
the already cached entry), making the initial Transients entry
'orphaned'. Subsequent requests to the initially cached resource now
stall, because after receiving the available Transients entry the Squid
worker keeps waiting for a notification from another worker, incorrectly
assuming that this worker (i.e., the entry writer) exists and is going
to fetch it. Since in reality there is no such writer, the transaction
hangs and is aborted after timeout.

In order to avoid this effect, we should not return such 'orphaned'
Transients entries. In other words, the caller may use a Transients
entry if and only if either this entry has been already anchored to a
cache or it has a writer (i.e., another worker is about to create that
cache soon).