Fixed getting entries with large headers from rock cache
Before this fix, when getting the entry from the rock cache, the
header parsing code could deal only with a single page at a time and
hence the maximum header size which could be parsed was 4K (the default
page size). Larger headers caused a parsing error with a level-1
message "Could not parse headers from on disk object" thus leaving
the reply unparsed. This invalid reply ( with "HTTP/1.1   0 Init"
status line was then sent back to the client).

Since the default maximum reply header size (as defined by
reply_header_max_size parameter) is 64K, we need to be able to parse
rock headers up to this value. Basically, we need to detect
whether a large header goes beyond the first slice boundary and parse
two (or more) slices in this case.