Removed the 'early messages' check from Debug:Enabled()
Since the early messages' maximum level is DBG_IMPORTANT (which
is now a default in Debug::Levels), we do not need an extra
condition for these messages to be allowed in Debug::Enabled().
All such messages are now passed to stderr and also stored if they
are 'early' in addition.