Removed a useless store_client::cmp_offset
This data member basically duplicates the offset stored in
store_client::copyInto and may be removed.  There was an arguable place
where it was used - the assertion on ONLYCONTIGUOUSREQUESTS flag is
provided.  However, this was added long ago (528b2c6) and seems like a
debugging code for the merged branch.