Reworked PeerSelectorTimeoutProcessor implementation
Use intrusive list of AsyncCalls instead of std::list of awaiting
PeerSelectors. This helped to simplify, avoiding storing additional data
such as the call position in the list and also improving performance
issues with inserting (and allocating) a new element into the list.
PeerSelector destructor still removes its pending call using a
linear search, but this inefficiency is not introduced by this PR: the
old event removal worked similarly.