Support asynchronous callbacks for non-blocking ACL checks, part 1
In this commit, we make sure that nobody stores a pointer to the
ACLChecklist object during a non-blocking ACL check. Storing such a
pointer will become unsafe because ACLChecklist will be gone by the time
the async callback fires.

I checked all nonBlockingCheck() callers. Those that did store the
checklist, did not really _use_ the stored checklist so I simply removed
that storage.

The only exception was PeerSelect that declared a BUG if it were to be
deleted before the non-blocking ACL check is completed, but that was not
a BUG -- the callback data is cbdata-protected so destroying PeerSelect
objects during a non-blocking check is safe and desirable. In fact, it
is deleting the checklist object under those conditions is very unsafe!
By removing that BUG-handling code, we actually fixed a bug. The
original reporting code was added in 348b203 because (in 1997, before
cbdata was introduced) the ACL check requestor was supposed to delete
the checklist. That responsibility was gone long time ago (I did not
check when).

This commit also removes a related stale comment.

No functionality changes intended.