Bug 5048: ResolvedPeers.cc:35: "found != paths_.end()" assertion
This bug was caused by a4d576d. When the destinations array was
searched for the 'retriable' path, connection object pointers were
compared. This, however, does not work for a persistent connection,
since 'destinations' contains a connection object other from the one
cached in fwdPconnPool and passed to ResolvedPeers::retryPath().
Instead, we should compare remote IP addresses, which are unique for
every 'destinations' element.