fixup: Simplify by removing timedOut()
When I used timedOut() in a sketch, I did not know that ping_data does
not know when the waiting ends. Yes, ping_data can extract that
information from the event processor, but the end result looks really
awkward: the event processor is asking ping_data to decide whether we
timed out and ping_data asks the event processor what the timeout was.

I also did not like the fact that ping_data cannot answer the "did you
timeout?" question in some of its states (without throwing). I was
worried that the question might be asked by the code that does not know
about this unnatural restriction.