fixup: Use assert() instead of Must()
I am not sure about this change because we do not yet have a good set of
criteria for when to use which checking method, unfortunately. Using
assert() is simpler and might offer slightly better performance. One of
the Must()s was executed in service context which would lack job
protections from transaction-specific exceptions. After checking that
there is currently high confidence in the related code, I used assert().