fixup: Break misleading start/stopPingWaiting() symmetry
While trying to document stopPingWaiting(), I discovered that we
mishandle stale handlePingTimeout() notifications (fixed in the previous
branch commit) and realized that start/stopPingWaiting() are naturally
asymmetrical. The start method is the right method to call to start
waiting for ping responses (the method will adjust state and set up the
corresponding timeout). However, that wait ends in three different ways:

* a timeout (where monitoring is ended without calling the method)
* a ping response (where the method is called and the state is changed)
* PeerSelector demise (where the method is called but the state is kept
  or, to be more precise, may/should be kept (for debugging/triage/etc.)

A similar conclusion may be reached by analyzing existing code that sets
PING_DONE -- some of that code does not want to end monitoring. Having a
method documented to enter the PING_DONE state that a lot of code should
avoid when entering that state is rather confusing.

In this commit, I am moving state management outside the method and
renaming the method to be focused exclusively on timeout monitoring.