fixup: Removed the pingWaiting() wrapper
The wrapper sometimes obscured the code intent because the surrounding
code was essentially enumerating PING_* values, and the wrapper hid
PING_WAITING. For example: PeerSelector::selectMore().

As often happens with a&&b wrappers, the negation was actually obscuring
the true meaning of the code because the reader was focusing on b while
a was equally important. For example: PeerSelector::handlePingTimeout().

This reversal also reduces (and clarifies) the diff.