Prevent from using stale Connection fields after retry
Elements of the 'destinations' array returned to the caller are
later modified, after opening that connection. If this connection
is retried afterwards, some stale connection fields may be accidentally
used somewhere - before the connection goes through ConnOpener::start()
for the second time. To avoid this, we now create and return to caller
a fresh candidate Connection.

Also simplified and polished.