ResolvedPeer is a (smart) Comm::Connection pointer, part 1
The branch changes so far increased my confidence in that the true
nature of the ResolvedPeer class is a (smart) Connection pointer (rather
than a Connection:index binding or pair).

In the ideal world, Squid would use some kind of [Resolved]PeerAddress
class to store the positive outcome of a peer selection process, but
Squid is already abusing the Connection class for holding the peer
address and other selection information. Fighting that bad design in the
middle of its users will only muddy the waters. We should reuse and bend
with the dual (or triple) nature of Comm::Connection instead.

I will adjust the code accordingly in 2+ steps. This first step focuses
on merging two FwdState Connection pointers into one and removing a few
ResolvedPeer methods that contradict the finalized class purpose. These
changes remove a lot of diff lines (compared to the base code) -- we are
only enhancing an existing Connection pointer, not adding another one.

The next step(s) will focus on polishing the ResolvedPeer API and
documentation.