Always return a Connection copy from ResolvedPeers
In 4299d17 we introduced a ResolvedPeerPath::dirty flag, so that
ResolvedPeers returned a path copy after it has been retried. This,
optimization introduced another problem because it is difficult to
make a 'pristine' path from 'dirty' by copying some fields. For example,
Connection::local address should be copied depending on a condition,
which is calculated only when 'destinations' array is created.

So we have to give up that optimization and  always return a path copy,
thus preserving pristine destinations in ResolvedPeers.