Do not throw in HandshakeParser::parseHello() if the buffer is empty
In the previous approach, the empty-buffer check was done in
HandshakeParser::parseHello(), however it makes sense to check this
earlier. I think we should not even start
ConnStateData::parseTlsHandshake() in this case (since there is nothing
to parse) and just read more data. On the other hand,
ConnStateData::afterClientRead() is able to resume tls parsing once
more data acquired.