Go back further
The alternative branch[1] goes forward "all the way" -- propagating
PeerConnectionPointer changes to all the connection opening, encrypting,
tunneling, etc. code. This branch does not go back far enough. The
current code highlights problems with my Connection hiding sketch in an
earlier review -- we cannot fully hide the open Connection inside
PeerConnection because HappyConnOpener and its Answer need access to it.

Without proper hiding, FwdState::serverConn meaning becomes too complex,
and the risk of misinterpreting/misapplying that object becomes too high
IMO. This change undoes merging of serverConn with
PeerConnectionPointer. Now, serverConn is again a simple ready-to-use
Comm::Connection pointer (or nil). A new, dedicated destinationReceipt
field is used for returnPath() management in FwdState.

HappyConnOpener changes should not be undone in a similar way. That
class or, to be precise, its Attempt components, do not _store_ the
FwdState::serverConn equivalent -- HappyConnOpener sends the answer and
stops as soon as the connection is found or established.

[1]SQUID-517-happy-eyeballs-disregard-viable-reforwarding-destinations-3