Polished naming and added a TODO
Here is the justification for the TODO:

The existing registerAbortCallback() callers always unlock the entry
-- today, this method signals the end of writer's ownership of the
StoreEntry. Calling unregisterAbortCallback() does not prevent an
already queued callback from being fired so any future hypothetical
unregisterAbortCallback() caller that does not end its StoreEntry
ownership will not be able to rely on that cancellation anyway -- the
method is called "unregister", but it is a rather poor mechanism for
ending notifications. Together with the surrounding code in callers,
this method is (implicitly) used for another purpose -- writing
ownership termination.