Check more preconditions before retrying TLS handshake failure with peer
Compared to bag14 631828a, the new conditions are "Squid is not shutting
down", "there is still time to connect", and "the client is still here".

Master commit 25d2603 adds a similar method with an additional check.
The master method checks whether the server connection is closed. That
check is in master for consistency or convenience sake: noConnections()
is called in many similar master TunnelStateData contexts. Master closes
the server connection before we checkRetry() so noConnections() in
master is equivalent to our "no client connection" condition.

Merged SQUID-478-reforward-connect-after-tls-handshake-failure-with-peer
...-bag14-2