fixup: Do not flush ostream manually
AFAICT, during regular use, std::ostream flushes as needed; manual
flushing is only necessary when we are about to yank the underlying
stream buffer from under the stream, without stream knowledge.
CollapsedForwarding::StatQueue() does not do anything like that.

I suspect that we should not flush PackableStream (before destroying it)
either, but the official code does that, and I would rather not add that
complicated question to this project, so I kept that flush() call in
StatQueues().