Add the HelperServerBase::canAcceptRequests method
... which checks if someone can send/write requests to the server,
to check that the server is not closing, the connection is open and
it is not closing.

Use this new method to check if the server is available to submit
new requets inside GetFirstAvailable/StatefulGetFirstAvailable functions,
instead of just checking the flags.closing flag. Checking the
flags.closing flag is not enough in these cases because the pipe may
closing because of an external reason.