Extend HelperServerBase::canAcceptRequests to cover more cases
The helper server can not accept requests if the helper is shuting down.
A statefull helper server can not accept requests if there is a pending
request
A statefull helper server can not accept requests if there is existing
reservation of helper which is not expired yet.

The HelperServerBase::canAcceptRequests is now a virtual method