Removed emphasis on ownership from LoadingParts
The concept of the (shared memory segment) owner is useful when there
are two different classes: the owner and the user. In case of
LoadingParts[Owner], the same class represents both (that is why we were
able to delete one of the LoadingParts* classes in this PR). There is
nothing unusual or special in that single representation -- the vast
majority of classes both own their data and use it -- so we should not
emphasize one aspect of such a class.

Also removed double allocation/leak of partsOwner in Rebuild::start().

Also restored explicit LoadingParts copying/moving prohibition. The
prohibition is no longer paranoid AFAICT -- the branch code was buggy
because it allowed LoadingParts objects to be copied while the class did
not support correct copying and moving.

Also restored important documentation of LoadingParts data members.