Changed more counters to int64_t (and renamed validatedCount)
The rationale here is the same as in branch commit ebd53bd "Polished
Progress counter types". This particular change started with the new
StoreRebuildData::validatedCount field which could already contain a sum
of entry and _slot_ counts across all cache_dirs. It was also compared
with dbEntryLimit which was initialized with int64_t entryLimitActual().

Again, we could add StoreRebuildData::validatedCount (now validations)
and other new data members/parameters with the wrong "int" type and then
change all related data members in a dedicated PR. I am not sure which
path is better, but I picked this one to avoid adding wrong types.