https_port...require-proxy-header connections are treated as intercepted
Disclosing that is important because several Squid configuration options
and logged warnings depend on whether the connection was intercepted.

Squid supports exactly three mutually exclusive proxy (port) roles:

* reverse -- The HTTP(S) client establishes an HTTP(S) connection with
  an origin server represented by this Squid listening port.

* interception -- The HTTP(S) client establishes an HTTP(S) connection
  with an origin server _not_ represented by this Squid listening port.

* forward -- The HTTP(S) client establishes an HTTP(S) connection with a
  proxy represented by this Squid listening port.

I am naming the second role "interception" here because "interception
proxy" is a common and even standard (RFC 3040) term. Both "intercept"
and "tproxy" proto_port modes configure an interception proxy. It is
unfortunate that the word "interception" explicitly matches the
"intercept" port option but not the "tproxy" one (where "t" stands for
"transparent" -- a synonym for an interception proxy).

It may feel counter-intuitive to call a regular haproxy-Squid TCP
connection "intercepted", but the term should be applied to the HTTP(s)
layer. In other words, "intercepted" is defined not by what was done to
the TCP connection to get its payload to Squid, but by how Squid handles
the accepted TCP connection and its payload.

Squid https_port supports all three roles: HTTPS reverse proxy, HTTPS
forward proxy, and an interception proxy handling 3rd party TLS
connections.

We can tell that the TLS connections this branch is dealing with were
intercepted because their SNI does not mention Squid, and their client
does not expect a Squid certificate. If a client wanted to use Squid as
an HTTPS proxy instead, that client would have named Squid in SNI and
expected a Squid server certificate. In the HTTPS forward proxy case,
any PROXY protocol prefix would have been _inside_ that client-to-Squid
TLS connection, preserving HTTPS proxy protections!

We do not want to add a new https_port mode keyword for this use case
because it is not necessary at this time. If we were to do so, it would
be a third keyword that enables the "interception" treatment of accepted
connections. The existing keywords are "intercept" and "tproxy".

N.B. In this context, HTTP(S) means "either HTTP or HTTPS". And HTTPS
includes both TLS and HTTP layers.