Fix adjusting CONNECT at step2 to copy all properties from origin request
The commit "Fix ssl bumping step2 to follow ..." builds a new request
with the adjusted SNI info, but all the other properties (headers, auth
info, notes etc) from step1 request are lost.
This patch instead of building a new HttpRequest object and try set its
parameters, clone the old one and replace the host name in URI and host
header, if the SNI info exist and if the CONNECT string has a IP address
and not a DNS name.
This is must be enough.

Changes:
  - Add a new ConnStateData::buildFakeRequest variant which takes as argument
    an HttpRequest object and builds a new ClientHttpRequest object.
  - Now the ConnStateData::startPeekAndSpliceStep2 instead of calling
    the fakeAConnectRequest or initiateTunneledRequest methods
    clones the original HttpRequest object, adjust it and use
    the new ConnStateData::buildFakeRequest method