Fix TlsBump3 related checks when generating client side certificates
Current code while the certificates are generated checks if peek and
splice is enabled at step1 to decide if should generate SSL/CTX objects
or configure already generated SSL/CTX objects (inside
ConnStateData::sslCrtdHandleReply and ConnStateData::getSslContextStart
methods). This is wrong because we may have decide to peek/stare
at step1 but the procedure break before we generate the SSLL/CTX
objects. This branch tried to fix this wrongly by checking if
we are enter the Step3 too.
This patch just checks if we have already generate SSL/CTX objects
and if yes reconfigure them with the generated certificates else
generate new SSL/CTX objects based on generated certificates.

Also this patch renames (and slightly fixes) the ConnStateData
doPeekAndSpliceStep() method to restartTlsNegotiation().
Actually this is what it does.