Make the Squid-specific TLS handling errors compatible with GnuTls
Use negative numbers assuming that the oficial errors are possitive
and also use only the first left bits set as values to keep
copmatibility with GnuTls bitwise X509 errors.
The GnuTls uses the first 20 right bits to define 20 different
certificate validation errors.

If the 32bits are not enough and required we can use 64bit number as
Security::ErrorCode.