Added protection from stale/unexpected IPC messages
The list of 'unexpected' cases does not include:

* external malicious messages, attacking Squid over IPC
* messages sent by misconfigured Squid instances
* messages sent from previous Squid instances
* messages from no-longer-running kid precesses
* wrong messages due to internal Squid bugs

The goal is to  protect the recipient from an answer to a question it
did not ask. This includes cases where the recipient has not asked the
same question and cases when the recipient has asked it. Given the above
exceptions, such unexpected answers may only arrive due to recipient
restarts.