fixup: Replaced an awkward QuestionerId(init) constructor
QuestionerId(true): There is probably only one best way for handling
these cases: The caller just needs to supply the ID of its own process.
MyQuestionerId() can now be used for that.

QuestionerId(false): We could use some bogus constructor parameter like
nullptr_t, but I noticed that all these cases are either in default
constructors or in constructors that are about to load the actual value
from the TypedMsgHdr. And the latter subcases are first constructing the
object with "default" values anyway (because we do not want to start
loading data members until we check the given message type). A default
constructor was a natural choice for both of these subcases.